pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hl7-to-gpap"
version = "0.1.0"
description = "Narrow MVP bridge between HL7 v2 ER7 messages and GPAP alarm messages"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "AGPL-3.0-only" }
authors = [
  { name = "PubInv contributors" }
]
dependencies = [
  "paho-mqtt>=2.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0"
]
hl7apy = [
  "hl7apy>=1.3.5"
]

[project.scripts]
hl7-to-gpap = "hl7_to_gpap.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]